pyproject.toml
[project]
name = "option-chain"
version = "0.1.0"
description = "Standalone Flask Option Chain Application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Flask==3.0.3",
    "python-dotenv==1.0.1",
    "websocket-client==1.8.0",
    "cachetools==5.3.3",
    "pytz==2024.1",
    "openalgo==1.0.39",
    "numpy==2.2.4",
    "numba==0.63.0b1",
    "llvmlite==0.46.0b1"
]

[tool.setuptools]
packages = ["utils"]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"
